fracture_detection/requirements.txt
torch
torchvision
numpy
opencv-python-headless
streamlit
matplotlib
scipy
Pillow
tqdm
pyyaml
ultralytics  
